third_party/blink/web_tests/virtual/webrtc-wpt-unified-plan/external/wpt/webrtc/RTCRtpSender-transport.https-expected.txt

This is a testharness.js-based test.
PASS RTCRtpSender.transport is null when unconnected
PASS RTCRtpSender/receiver.transport has a value when connected
PASS RTCRtpSender/receiver.transport at the right time, with bundle policy balanced
PASS RTCRtpSender/receiver.transport at the right time, with bundle policy max-bundle
PASS RTCRtpSender/receiver.transport at the right time, with bundle policy max-compat
Harness: the test ran to completion.
